requirements.txt
streamlit
torch
torchvision
timm
numpy
pillow
opencv-python-headless